Added a Game State System
Created a state class where we can change the state of our game which allows us to change how the game updates and renders. This version is highly experimental and currently has a CIRCULAR DEPENDANCY between world, game and ActivePlayState. another state must be created to check if state switching functions as intended. DO NOT MERGE TO ANY BRANCH AS THIS IS NOT A STABLE VERSION